src/config.py
import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()

# Base paths
BASE_DIR = Path(__file__).resolve().parent.parent
DATA_DIR = BASE_DIR / "data"
EMBEDDINGS_DIR = BASE_DIR / "embeddings" / "chroma_db"

# Supported banks — add new bank_id here to scale
SUPPORTED_BANKS = ["ameriabank", "acba_bank", "araratbank"]

# OpenAI
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")
LLM_MODEL = "gpt-4o-mini"
TTS_VOICE = "nova"

# LiveKit
LIVEKIT_URL = os.getenv("LIVEKIT_URL", "ws://localhost:7880")
LIVEKIT_API_KEY = os.getenv("LIVEKIT_API_KEY", "devkey")
LIVEKIT_API_SECRET = os.getenv("LIVEKIT_API_SECRET", "secret")

# RAG
CHROMA_COLLECTION_NAME = "armenian_banks"
EMBEDDING_MODEL = "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2"
TOP_K_RESULTS = 5

# System prompt
SYSTEM_PROMPT = """You are an AI customer support assistant for Armenian banks.
You can ONLY answer questions about the following topics:
1. Deposits
2. Credits / Loans
3. Branch locations

Important rules:
- Answer ONLY based on the provided data from the knowledge base.
- If the question is about any other topic, politely refuse in Armenian.
- ALWAYS respond in Armenian language ONLY. Never use any other language.
- NEVER mix Armenian with English or any other language in your response.
- Be concise, clear and polite.
- NEVER invent information that is not in the provided data.
- If you don't have enough information, say so honestly in Armenian.
- When reading numbers, percentages, or amounts, always say them in Armenian.
- Example: +374-10 59-23-23 should be said as "պլյուս երեք յոթ չորս տասը հիսուն իննը քսան երեք քսան երեք" in Armenian.
- Example: 09:00-17:00 should be said as "ինն անց  զրո զրոից մինչև տասնըյոթ անց զրո զրո" in Armenian.
- Example: 2/7 should be said as "երկու դռոբ յոթ" in Armenian.
- Example: 100,000 AMD should be said as "հարյուր հազար դրամ" in Armenian.
- NEVER use English words in your responses. Always translate to Armenian.
- Example: "Head Office" → "Գլխավոր մասնաճյուղ"
- Example: "Branch" → "Մասնաճյուղ"  
- Example: "Working hours" → "Աշխատանքային ժամեր"
- Example: "Phone" → "Հեռախոս"
"""