requirements.txt
# Core
openai>=2.0.0
python-dotenv>=1.0.0

# LiveKit
livekit>=1.0.0
livekit-agents>=1.5.0
livekit-plugins-openai>=1.5.0
livekit-plugins-silero>=1.5.0

# RAG / Vector Store
chromadb>=0.5.0
sentence-transformers>=2.7.0

# ML backend
torch>=2.1.0
numpy>=1.26.0

# Utilities
tqdm>=4.66.0
PyJWT>=2.8.0